+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fixed
+- Use webpack plugin versions compatible with webpack@4
+- Fix typo in .babelrc
+
 ## [1.0.0] - 2018-04-06
 ### Added
 - Prettier code formatter